requirements/requirements-dev.txt
# dev work/tooling
ipdb
click

# graphing libs
matplotlib

# releasing
twine

jupyterlab
jupyterlab-dash
jupyter_dash
nbstripout
plotly

